Block {DiGGer}	R Documentation

DiGGer Block object constructor.

Description

Create a Block object component of a DiGGer objective.

Usage

Block(rowsInBlock = 1, columnsInBlock = 1, blockVarianceRatio = 0.1)

Arguments

rowsInBlock
Number of rows in the template block for this Block object.

columnsInBlock
Number of columns in the template block for this Block object.

blockVarianceRatio
Variance component ratio for this block. A random block has a positive component otherwise the block is a fixed block.

Details

Blocks, fixed or random, are component objects in a DiGGer search Objective. Blocks tile down and across from Row 1 and Column 1 of a matrix representation of the design. The final block in a column or row may be truncated to conform to the shape of the design. The default Block() will add a "nugget" variance component when added to a DiGGer Objective.

Value

.nrpb
Number of rows in the template block.

.ncpb
Number of columns in the template block.

.gamma
The variance component ratio associated with the blocking factor.

Author(s)

Neil Coombes

[Package DiGGer version 1.0.5 Index]