Correlation {DiGGer}	R Documentation

DiGGer Correlation object constructor.

Description

Create a Correlation object component of a DiGGer objective.

Usage

Correlation(rowModel = "AR", rowParameter = 0.5, columnModel = "AR",
columnParameter = 0.5, spatialVarianceRatio = 1) 

Arguments

rowModel
The correlation pattern between rows may be "AR" AutoRegressive, "MA" Moving Average, or "ID" no correlation.

rowParameter
Valid parameter values p, for "AR" -1<p<1, for "MA" -0.5<p<0.5 and for "ID" p=0.

columnModel
As for rowModel.

columnParameter
As for rowParameter.

spatialVarianceRatio
The variance component ratio associated with the correlation between experimental units.

Details

The correlation in an "AR" process with parameter p decreases with separation according to p^|i-j| where |i-j| is the separation between rows or columns. The correlation in an "MA" process is p for adjacent rows or columns and 0 otherwise. There is no correlation between rows or columns for an identity, "ID", model.

Value

.rmod
Row model of correlation object.

.rhor
Correlation parameter for rows.

.cmod
Column model of correlation object.

.rhoc
Correlation parameter for columns.

.gamma
Variance component ratio of correlation object.

Author(s)

Neil Coombes

[Package DiGGer version 1.0.5 Index]