DiGGer {DiGGer}	R Documentation

Constructor function for a DiGGer design.

Description

DiGGer initialises a DiGGer object which may be modified to create a search for producing efficient experimental designs.

Usage

DiGGer(numberOfTreatments = 18, rowsInDesign = 18, columnsInDesign = 3,
rowsInReplicate = NULL, columnsInReplicate = NULL,
treatName = NULL, treatNumber = NULL,
treatRepPerRep = NULL, treatGroup = NULL,
treatType = "FIXED", treatVarianceRatio = -1, initialDesign = NULL,
initialSwap = NULL, rngSeeds = NULL, rngState = NULL)

Arguments

numberOfTreatments
The number of treatments in the design.

rowsInDesign
The number of rows in the design.

columnsInDesign
The number of columns in the design.

rowsInReplicate
The number or rows in the template replicate block.

columnsInReplicate
The number of columns in the template replicate block.

treatName
Vector of treatment names to be associated with the design numbers.

treatNumber
Vector of treatment numbers to be associated with the design numbers.

treatRepPerRep
Vector of replication levels for each treatment within each replicate template block.

treatGroup
Vector of group codes (up to 200 distinct values) associated with treatments. Group codes may be used to modify the A-efficiency measure in the optimisation.

treatType
By default treatments are of "FIXED" type. Treatments of "RANDOM" type require a treatVarianceRatio.

treatVarianceRatio
A positive value is required for "RANDOM" treatment types, otherwise treatments are "FIXED".

initialDesign
A matrix rowsInDesign by columnsInDesign giving design numbers in the initial design.

initialSwap
A matrix rowsInDesign by columnsInDesign of swap codes. Only plots with the same swap code may have treatment interchanges during the DiGGer search.

rngSeeds
Seeds c(s1,s2) to control the DiGGer search. s1 must be in the range [0,31328], s2 must be in the range [0,30081].

rngState
Current state of the random number generator.

Details

DiGGer constructs a framework for a DiGGer search for efficient experimental designs. The random number generator is initialised. Treatment information is used to create a treatment data frame.

DiGGer establishes an initialDesign and an initialSwap matrix. If no initialDesign is supplied the parameters, numberOfTreatments, rowsInDesign, columnsInDesign, rowsInReplicate, columnsInReplicate and treatRepPerRep, are used to generate an initial design. If no initialSwap matrix is supplied a matrix of 1s is set up to allow all treatment plot interchanges within a replicate.

Wrapper functions are applied to the DiGGer object to create searches for common design types. Alternatively constructor functions Block, Correlation, Objective and Phase may be used to modify the search framework to create a set of search instructions.

Value

A DiGGer search object for a null search.

idsgn
An initial design matrix of design numbers.

iswap
An initial matrix of swap codes. Swaps are possible between units with the same swap code.

treatment
A treatment data frame with columns ID, ENTRY, Repeats and Group.

ddphase
An empty list to be populated with search instructions.

Author(s)

Neil Coombes

Examples

# Example of initialisation for 21 treatments
# arranged in 21 rows by 3 columns
# with replicates 21 rows by 1 column
d21 <- DiGGer(21, 21,3, 21,1)

[Package DiGGer version 1.0.5 Index]