DiGGer-package {DiGGer}	R Documentation

DiGGer experimental design generator.

Description

DiGGer generates A-efficient experimental designs under specified correlation and/or blocking.

Details

Package: 	 DiGGer

Type: 	 Package

Version: 	 0.3-0

Date: 	 2016-01-01

License: 	 NSW DPI Freeware Licence

The DiGGer function creates a DiGGer search object which may be modified before being run using the run method.

Author(s)

Neil Coombes

Maintainer: Neil Coombes <neil.coombes@dpi.nsw.gov.au>

References

Coombes, N.E. (2002) The Reactive Tabu Search for Efficient Correlated Experimental Designs. PhD Thesis, Liverpool John Moores University.

Examples

#
# Example of search set up for 21 treatments
# arranged in 21 rows by 3 columns
# with replicates 21 rows by 1 column
d21 <- corDiGGer(21, 21,3, 21,1)
d21 <- run(d21)
dmatrix <- getDesign(d21)
dframe <- d21$dlist

[Package DiGGer version 1.0.5 Index]