Objective {DiGGer}	R Documentation

Constructor function for a DiGGer Objective.

Description

Construct a DiGGer Objective for use in a DiGGer search.

Usage

Objective(weight = 1, linearCovariate = "NONE", numberOfBlocks = 0,
blocks = NULL, corr = Correlation())

Arguments

weight
The weight to scale the A-measure for this objective.

linearCovariate
Code for the linear covariate to include in the objective: "NONE", "ROWS", "COLUMNS" or "BOTH".

numberOfBlocks
Number of block factors in the Objective.

blocks
List of Block() objects.

corr
A Correlation() object.

Details

A DiGGer Objective() creates a covariance between experimental units which is used to calculate the efficiency of the design using an A-measure criterion. The efficiency measure is scaled by the weight parameter and may be summed with efficiency measures of other objectives.

A DiGGer Objective may include fixed covariates based on row or column number and fixed block effects. Random components will include a component for variability between experimental units which may have separable correlation processes, auto-regressive "AR" or moving average "MA" between rows and columns. Random blocks may also be present.

Value

.weight
weight used to scale the A-measure associated with this objective.

.lincov
Code for the linear covariates, "NONE", "ROWS", "COLUMNS" or "BOTH".

.nblk
Number of block factors.

.blk
List of Block() factors.

.corri
A Correlation() object.

Author(s)

Neil Coombes

[Package DiGGer version 1.0.5 Index]