Phase {DiGGer}	R Documentation

DiGGer Phase object constructor function.

Description

Create a Phase object component of a DiGGer search.

Usage

Phase(rowsInSwapBlock = 1, columnsInSwapBlock = 1,
rowsInCorrelationBlock = 1, columnsInCorrelationBlock = 1, aType =
"A++", targetAValue = 0, maxInterchanges = 1e+05, searchIntensity = 100,
numberOfObjectives = 1, objectives = list(Objective()))

Arguments

rowsInSwapBlock
Number of rows in the Swap Block template.

columnsInSwapBlock
Number of columns in the Swap Block template.

rowsInCorrelationBlock
Number of rows in the blocks of correlated error.

columnsInCorrelationBlock
Number of columns in the blocks of correlated error.

aType
The A-measure code, "A++", "Agg", "A22", "A11", "A1+", "Aa2" or "Aaa".

targetAValue
A-measure stopping rule for this Phase.

maxInterchanges
Maximum number of treatment interchanges stopping rule for this Phase.

searchIntensity
Percentage of interchanges to be considered for non-improving treatment interchange.

numberOfObjectives
Number of objectives in this search Phase.

objectives
List of DiGGer Objective() objects to be optimised.

Details

A DiGGer search Phase is the framework for a call to the DiGGer search program. The Phase object is a component of a DiGGer object. When the DiGGer search is run each Phase optimises the current design subject to the A-measure aType measured on the Objectives specified. A DiGGer search may have more than one Phase. After a search Phase the optimised design is carried forward to the next Phase of the search.

In this way a DiGGer search involves progressive refinement of a design.

Swap block dimensions should be adjusted at the start of each search Phase so that objectives already optimised are not compromised. This will mean that swap blocks should not be bigger than the smallest block in a completed search Phase.

The blocks of correlated error defined by rowsInCorrelationBlock and columnsInCorrelationBlock must tile evenly over the design dimensions.

Value

.nrbs
Number of rows in the Swap Block template.

.ncbs
Number of columns in the Swap Block template.

.nrbv
Number of rows in the blocks of correlated error.

.ncbv
Number of columns in the blocks of correlated error.

.atype
A-measure code.

.targa
Target A-measure.

.maxint
Maximum interchanges to test.

.srchint
Search intensity.

.nobj
Number of objectives.

.objve
List of DiGGer Objectives.

Author(s)

Neil Coombes

[Package DiGGer version 1.0.5 Index]