addObjective {DiGGer}	R Documentation

Add an Objective to a DiGGer search.

Description

Add an Objective to a search phase in a DiGGer search.

Usage

## S3 method for class 'DiGGer'
addObjective(this, phaseNo = 1, newObjective =
Objective(), ...)
## S3 method for class 'Phase'
addObjective(this, newObjective = Objective(weight = 1.00,
linearCovariate = "NONE", numberOfBlocks = 0, blocks=NULL,
corr=Correlation(), ...), ...)

Arguments

this
The DiGGer component object to be modified.

phaseNo
The Phase of the DiGGer search to be modified.

newObjective
The Objective() to be added. The default objective has autoregressive (AR(0.5)) spatial correlations between rows and between columns with no blocks.

...
Not used.

Author(s)

Neil Coombes

See Also

Objective.

[Package DiGGer version 1.0.5 Index]