addPhase {DiGGer}	R Documentation

Add a Phase object to a DiGGer search.

Description

Add a Phase() object to a DiGGer search.

Usage

## S3 method for class 'DiGGer'
addPhase(this, newphase = Phase(), ...)

Arguments

this
The DiGGer object to be modified.

newphase
A DiGGer Phase object. The default Phase() object must be modified by specifying swap block parameters, rowsInSwapBlock, columnsInSwapBlock, and correlation block parameters, rowsInCorrelationBlock, columnsInCorrelationBlock which all have default values of 1.

...
Not used.

Warning

Failure to modify the default Phase() settings will result in errors in swap block dimensions and correlation block dimensions.

Author(s)

Neil Coombes

See Also

Phase.

[Package DiGGer version 1.0.5 Index]