createFactorialDF {DiGGer}	R Documentation

Create a factorial treatment data frame.

Description

Create a factorial treatment data frame with factor levels and replication levels.

Usage

createFactorialDF(nlevelsVector = NULL, treatName = NULL, treatRepPerRep
= NULL.)

Arguments

nlevelsVector
A vector of the number of levels in each factor.

treatName
A list of named vectors. Each vector may contain labels associated with the levels of each factor.

treatRepPerRep
Repeats per replicate of each factor combination. If not given each treatment combination has a repeats value of 1.

Details

Creates a treatment data frame with numeric levels and factor labels for factors with replication levels per replicate.

Author(s)

Neil Coombes

[Package DiGGer version 1.0.5 Index]