deleteBlock {DiGGer}	R Documentation

Delete a Block from an objective in a DiGGer search

Description

Delete a Block from an objective in a DiGGer search.

Usage

## S3 method for class 'DiGGer'
deleteBlock(this, phaseNo = 1, objectiveNo = 1, blockNo = 1, ...)
## S3 method for class 'Phase'
deleteBlock(this, objectiveNo = 1, blockNo = 1, ...)
## S3 method for class 'Objective'
deleteBlock(this, blockNo = 1, ...)

Arguments

this
The DiGGer object to be modified

phaseNo
The Phase of the DiGGer search to be modified.

objectiveNo
The Objective to be modified.

blockNo
The block to be deleted from the objective.

...
Not used.

Author(s)

Neil Coombes

[Package DiGGer version 1.0.5 Index]