des.embedopr01 {DiGGer}	R Documentation

Function to generate an embedded p-rep design

Description

Generate an optimised embedded p-rep design within an optimised replicated design.

Usage

des.embedopr01(NumberOfTreatments, RowsInDesign, ColumnsInDesign,
RowsInReplicate, ColumnsInReplicate, B2D=TRUE, RowsInEmbeddedBlock,
ColumnsInEmbeddedBlock, InitialDesign=NULL, ...)

Arguments

NumberOfTreatments
The number of distinct treatments in the design.

RowsInDesign
The number of rows in the design.

ColumnsInDesign
The number of columns in the design.

RowsInReplicate
The number of rows in each replicate block.

ColumnsInReplicate
The number of columns in each replicate block.

B2D
Logical call for blocks in two directions or a dimension pair c(r,c) defining the intersection blocks of two row block and column block structures.

RowsInEmbeddedBlock
The number of rows in the embedded block.

ColumnsInEmbeddedBlock
The number of columns in the embedded block.

InitialDesign
A matrix RowsInDesign by ColumnsInDesign giving design numbers in the initial design.

...
Other arguments to the DiGGer function.

Details

des.embedopr01 creates an optimised partially replicated design within a fully replicated design. The function aims to get a replicate of treatments and some duplicates within the embedded block while maintaining the integrity of the replicated design. The embedded partially replicated design is optimised respecting replicate boundaries of the full design. This design is kept fixed while the fully replicated design is optimised around it.

Value

A DiGGer design object with the specifications for and results of a search. A graph of the design showing the embedded block and the replicated treatments is also produced.

Author(s)

Neil Coombes

Examples

# design for 44 entries [44x3] with replicates [44x1]
# and embedded block [25x2].
d44x3k25 <- des.embedopr01(NumberOfTreatments=44,
                     RowsInDesign=44,
                     ColumnsInDesign=3,
                     RowsInReplicate=44,
                     ColumnsInReplicate=1,
                     B2D=TRUE,
                     RowsInEmbeddedBlock=25,
                     ColumnsInEmbeddedBlock=2)
View(d44x3k25$fdsgn$dlist)

[Package DiGGer version 1.0.5 Index]