desPlot {DiGGer}	R Documentation

Plot a design as a matrix

Description

Plot a matrix representing an experimental design to the current graphics device.

Usage

desPlot(dsgn, trts = NULL, rprop = 1, cprop = 1, label = TRUE, plotchar
= NULL, plotbndry = TRUE, chtdiv = 2, bdef = NULL, bcol = 1, bwd = 2,
rotate = FALSE, new = TRUE, cstr = "Range", rstr = "Row", rlab = TRUE,
clab = TRUE, ...)

Arguments

dsgn
A matrix representation of an experimental design.

trts
Specific treatment labels to be plotted.

rprop
Proportion of the row boundary of an experimental unit to plot.

cprop
Proportion of the column boundary of an experimental unit to plot.

label
If TRUE print a label for all treatments or specific treatments listed in trts. If FALSE labels are not printed.

plotchar
A single text character or plotting symbol to be used as the label.

plotbndry
If plotbndry is TRUE, a polygon is used to plot the experimental unit.

chtdiv
A reduction factor for the plotting symbol.

bdef
A matrix rx2. Each row defines the rows and columns in a block template to be plotted.

bcol
The colour used for the block boundary.

bwd
The width of the block boundary to be plotted.

rotate
If TRUE the matrix is rotated 90 degrees for plotting.

new
If TRUE a new plot is produced, if FALSE the current plot is added to.

cstr
String to label Columns of the matrix.

rstr
String to label Rows of the matrix.

rlab
If TRUE rows of the design are labelled.

clab
If TRUE columns of the design are labelled.

...
Other arguments to polygon used for plotting experimental units in the design.

Details

des.plot may be used to build a graphical representation of a matrix, highlighting the position of certain treatments in the design and the blocking factors used in the design.

Value

A plot is produced.

Author(s)

Neil Coombes

[Package DiGGer version 1.0.5 Index]