desTab {DiGGer}	R Documentation

Tabulate replication levels of treatments in blocks.

Description

Tabulate replication levels of treatments or index factor within each block of a design.

Usage

desTab(designMat, rowsInBlock, columnsInBlock, groupCode=NULL)

Arguments

designMat
A matrix representation of a design.

rowsInBlock
Number of rows in the block template.

columnsInBlock
Number of columns in the block template.

groupCode
An index of codes associated with the design numbers to be tabulated.

Details

Table of frequencies of treatments or index codes within each block.

Author(s)

Neil Coombes

[Package DiGGer version 1.0.5 Index]