errcodes {DiGGer}	R Documentation

DiGGer error message texts.

Description

errcodes contains the texts for error messages that may be generated by a call to the DiGGer design generation program. getError() operates on a DiGGer object to print the error message.

Usage

data(errcodes)

[Package DiGGer version 1.0.5 Index]