facDiGGer {DiGGer}	R Documentation

Function to generate a DiGGer search for a factorial design

Description

Generate a DiGGer search for designs with factorial treatments.

Usage

facDiGGer(factorNames = NULL, rowsInDesign = NULL, columnsInDesign =
NULL, rowsInReplicate = NULL, columnsInReplicate = NULL, mainPlotSizes =
NULL, treatDataFrame = DF, treatRepColumn = "Repeats",
blockSequence = NULL, objectiveWeight = c(0.8, 0.2), searchIntensity =
100, chequerboard = FALSE, spatial = TRUE, rowColumn = TRUE,
designLayoutTemplate = NULL, maxInterchanges = 100000, rngSeeds =
NULL, rngState = NULL)

Arguments

factorNames
Text vector of factor names.

rowsInDesign
Number of rows in the design.

columnsInDesign
Number of columns in the design.

rowsInReplicate
Number or rows in each replicate block.

columnsInReplicate
Number of columns in each replicate block.

mainPlotSizes
A list of dimension pairs indicating main plot sizes. Randomised factorial designs have main plot sizes of c(1,1) whereas split plot designs have the dimensions of the main plots.

treatDataFrame
Data frame with treatment factors and repeats per replicate.

treatRepColumn
Text name of the column in the treatDataFrame with treatment repeats information.

blockSequence
List of dimension pairs of blocks to be optimised in sequence.

objectiveWeight
Relative weights for multiple objectives – limited use.

searchIntensity
Percentage of possible interchanges to consider for non-improving interchanges – limited use.

chequerboard
Used with 2 level factors. If FALSE the spatial setting is not used with this factor, otherwise spatial correlation will favour a chequerboard arrangement of the factor levels.

spatial
If TRUE separable AR processes between rows and between columns are specified in the final DiGGer search phase for each factor.

rowColumn
If TRUE random rows, random columns and random error are specified in the final DiGGer search phase for each factor.

designLayoutTemplate
An optional matrix showing replicate numbers with missing plots coded as 0.

maxInterchanges
Maximum number of interchanges used in each search phase.

rngSeeds
Seeds c(s1,s2) to control the DiGGer search. s1 must be in the range [0,31328], s2 must be in the range [0,30081].

rngState
Current state of the random number generator.

Details

facDiGGer is a one factor at a time search which optimises the first factor across the whole design before considering each subsequent factor in order. As each factor is added to the design the number of treatments in the design is increased and the levels of the latest factor added are used as a group code. The last factor added may have limited randomisation.

Split-plot designs are created by creating main plots and expanding them to the appropriate size before adding subsequent factors.

The createFactorialDF function is provided to help create factorial treatment data frames.

Value

A DiGGer object with the sequentially optimised design.

Author(s)

Neil Coombes

References

Coombes, N.E. (2002) The Reactive Tabu Search for Efficient Correlated Experimental Designs. PhD Thesis, Liverpool John Moores University.

Examples

# 3x3x3 factorial
# blocked to give replicates in two directions
DF27 <- createFactorialDF(c(3,3,3))
test27 <- facDiGGer(c("F1","F2","F3"),
                    27,3, 27,1, treatDataFrame = DF27,
                    rngSeeds = c(2468, 1357),
                    blockSequence = list(c(9,1)))

[Package DiGGer version 1.0.5 Index]