findblks {DiGGer}	R Documentation

Generate block codes.

Description

findblks uses the design dimensions and the block dimensions to generate a block code for each experimental unit in a design.

Usage

findblks(nrbd, ncbd, nrpb, ncpb)

Arguments

nrbd
Number of rows in the design.

ncbd
Number of columns in the design.

nrpb
Number of rows in the template block.

ncpb
Number of columns in the template block.

Details

Blocks are tiled from row 1, column 1 down and across the design with blocks numbered in vec order. The final block in a row or column may contain fewer units.

Value

A matrix of block identifying numbers.

Author(s)

Neil Coombes

[Package DiGGer version 1.0.5 Index]