getBlock {DiGGer}	R Documentation

Get a Block object from a DiGGer search object.

Description

Get a copy of a Block object from a DiGGer search object.

Usage

## S3 method for class 'DiGGer'
getBlock(this, phaseNo = 1, objectiveNo = 1, blockNo =
1, allBlocks=FALSE, ...)
## S3 method for class 'Phase'
getBlock(this, objectiveNo = 1, blockNo = 1, ...)
## S3 method for class 'Objective'
getBlock(this, blockNo = 1, ...)

Arguments

this
A DiGGer, component object from a DiGGer search.

phaseNo
The Phase to be queried.

objectiveNo
The Objective to be queried.

blockNo
The Block to be returned.

allBlocks
If TRUE a dataframe of all blocks in the design is returned.

...
Not used.

Warning

Modification of this Block object will also modify the copy in its original context, Objective, Phase, DiGGer. A new Block() assignment to this object will not affect the original context.

Author(s)

Neil Coombes

[Package DiGGer version 1.0.5 Index]