getConcurrence.DiGGer {DiGGer}	R Documentation

Get a Treatment-Block or Block-Treatment concurrence table summary.

Description

Get a Treatment-Block or Block-Treatment concurrence table summary for a DiGGer experimental design.

Usage

## S3 method for class 'DiGGer'
getConcurrence(this, phaseNo = 1, objectiveNo = 1, blockNo = 1,
type = "Treatment-Block", ...)

Arguments

this
A DiGGer design object.

phaseNo
The Phase containing the Block object of interest.

objectiveNo
The Objective containing the Block object of interest.

blockNo
The Block object to be used in the summary.

type
Either "Treatment-Block" or "Block-Treatment" concurrences to be tabulated.

...
Not used.

Details

getConcurrence uses Ntb=table(design,block) to tabulate treatment concurrences in blocks then produces the summary table of the lower triangular matrix Ntb%*%t(Ntb) for the concurrences of pairs of treatments within blocks, or t(Ntb)%*%Ntb for the number of treatments in common between pairs of blocks.

Value

Table of concurrences

Author(s)

Neil Coombes

[Package DiGGer version 1.0.5 Index]