getCorrelation {DiGGer}	R Documentation

Get a Correlation object from a DiGGer search.

Description

Get a Correlation object from an objective in a DiGGer search.

Usage

## S3 method for class 'DiGGer'
getCorrelation(this, phaseNo = 1, objectiveNo = 1, ...)
## S3 method for class 'Phase'
getCorrelation(this, objectiveNo = 1, ...)
## S3 method for class 'Objective'
getCorrelation(this, ...)

Arguments

this
A component object from a DiGGer search.

phaseNo
The Phase to be accessed.

objectiveNo
The Objective to be accessed.

...
Not used.

Warning

Modification of this Correlation object will also modify the copy in its DiGGer context. A new Correlation() assignment to this object will not affect the DiGGer Objective correlation.

Author(s)

Neil Coombes

[Package DiGGer version 1.0.5 Index]