getCorrelationBlock {DiGGer}	R Documentation

Show the dimensions of a correlation block in a DiGGer search.

Description

Show the dimensions of a correlation block in a DiGGer search.

Usage

## S3 method for class 'DiGGer'
getCorrelationBlock(this, phaseNo = 1, ...)
## S3 method for class 'Phase'
getCorrelationBlock(this, ...)

Arguments

this
A DiGGer or Phase object of a DiGGer search.

phaseNo
The Phase of the DiGGer search to be accessed.

...
Not used.

Author(s)

Neil Coombes

[Package DiGGer version 1.0.5 Index]