getidsgn {DiGGer}	R Documentation

Generate an initial design for a DiGGer search.

Description

Generate an initial design for a DiGGer search.

Usage

getidsgn(ntrt, trep = NULL, nrbd, ncbd, nrbr, ncbr)

Arguments

ntrt
Number of distinct treatments in the design.

trep
Vector of treatment replications for each treatment in a design replicate.

nrbd
Number of rows in the design.

ncbd
Number of columns in the design.

nrbr
Number of rows in each replicate.

ncbr
Number of columns in each replicate.

Value

A matrix of design numbers matching the layout and replication levels specified.

Author(s)

Neil Coombes

[Package DiGGer version 1.0.5 Index]