ibDiGGer {DiGGer}	R Documentation

Function to search for incomplete block designs using DiGGer.

Description

ibDiGGer modifies a DiGGer object to search for an incomplete block design.

Usage

ibDiGGer(numberOfTreatments, rowsInDesign, columnsInDesign,
rowsInReplicate = NULL, columnsInReplicate = NULL,
rowsInBlock, columnsInBlock, fixedBlocks = FALSE,
blockGamma = 1.0, aType = NULL,
targetAValue = 0, maxInterchanges = 1e+05,
searchIntensity =100, runSearch = TRUE,
rngSeeds = NULL, rngState = NULL, treatName = NULL,
treatNumber = NULL, treatRepPerRep = NULL, treatGroup = NULL,
initialDesign = NULL, initialSwap = NULL)

Arguments

numberOfTreatments
The number of treatments in the design.

rowsInDesign
The number of rows in the design.

columnsInDesign
The number of columns in the design.

rowsInReplicate
The number or rows in the template replicate block.

columnsInReplicate
The number of columns in the template replicate block.

rowsInBlock
The number of rows in each block.

columnsInBlock
The number of columns in each block.

fixedBlocks
Logical value whether blocks are treated as fixed.

blockGamma
Variance component assigned to blocks.

aType
A DiGGer A-measure type: "A++", "Agg", "A22", "A11", "A1+", "Aa2" or "Aaa". The default is "A++" for equally replicated treatments, "Agg" for unequal replication.

targetAValue
The search stops in each search phase when the A value is below this value.

maxInterchanges
Number of treatment interchanges to test.

searchIntensity
Percentage of possible interchanges to consider for non-improving interchanges.

runSearch
Logical value, whether to run the search immediately.

rngSeeds
Seeds c(s1,s2) to control the DiGGer search. s1 must be in the range [0,31328], s2 must be in the range [0,30081].

rngState
Current state of the random number generator.

treatName
Vector of treatment names to be associated with the design numbers.

treatNumber
Vector of treatment numbers to be associated with the design numbers.

treatRepPerRep
Vector of replication levels for each treatment within each replicate template block.

treatGroup
Vector of group codes (up to 200 distinct values) associated with treatments. Group codes may be used to modify the A-efficiency measure in the optimisation.

initialDesign
A matrix rowsInDesign by columnsInDesign giving design numbers in the initial design.

initialSwap
A dimension pair or a matrix rowsInDesign by columnsInDesign of swap codes. Only plots with the same swap code may have treatment interchanges during the DiGGer search.

Details

ibDiGGer constructs a framework for a DiGGer incomplete block design search. Blocks are treated as random by default. The search may fail with fixed blocks if the blocks become disconnected. When runSearch = TRUE the search is run when ibDiGGer is called otherwise the run function may be called to run the search.

If there is more than one treatGroup code the search may optimise comparisons between treatments in different treatment groups, aType="Agg". If aType=NULL the default code for equally replicated treatments is "A++" and for unequal replication it is "Agg".

Value

A DiGGer search object.

ddphase
A list with an element for each phase of the search. An element is populated with seeds, design, aMeasures and lastImprovement as each search phase is completed.

dlist
A dataframe of the design after a search has been run().

Author(s)

Neil Coombes

References

Coombes, N.E. (2002) The Reactive Tabu Search for Efficient Correlated Experimental Designs. PhD Thesis, Liverpool John Moores University.

Examples

# Example of search run for 12 treatments
# in 3 3-row by 4-column replicates
# and default nested rows and columns
rc12 <- rcDiGGer(12, 3,12, 3,4)

[Package DiGGer version 1.0.5 Index]