nurseryDiGGer {DiGGer}	R Documentation

Function to generate a DiGGer search replicated design with high check treatment frequencies

Description

Generate a DiGGer search for designs with replicated treatments and some high frequency check treatments.

Usage

nurseryDiGGer(numberOfTreatments, rowsInDesign, columnsInDesign,
rowsInReplicate = NULL, columnsInReplicate = NULL,
blockSequence = NULL, betweenRowCorrModel = "AR", betweenRowCorr = 0.5,
betweenColumnCorrModel = "AR", betweenColumnnCorr = 0.5, treatName =
NULL, treatNumber = NULL, treatRepPerRep = NULL, treatGroup = NULL,
maxInterchanges = 20000, checkGroup = NULL, runSearch = FALSE, rngSeeds =
NULL)

Arguments

numberOfTreatments
Number of distinct treatments in the design.

rowsInDesign
Number of rows in the design.

columnsInDesign
Number of columns in the design.

rowsInReplicate
Number of rows in each replicate.

columnsInReplicate
Number of columns in each replicate.

blockSequence
List of dimension pairs of blocks to be optimised in sequence.

betweenRowCorrModel
The correlation pattern between rows may be "AR" AutoRegressive, "MA" Moving Average, or "ID" no correlation.

betweenRowCorr
Valid parameter values p, for "AR" -1<p<1, for "MA" -0.5<p<0.5 and for "ID" p=0.

betweenColumnCorrModel
As for betweenRowCorrModel.

betweenColumnCorr
As for betweenRowCorr.

treatName
Vector of texts giving the names of the treatment.

treatNumber
Vector of numbers associated with the treatments.

treatRepPerRep
Vector of replication levels for each treatment over the whole design.

treatGroup
Vector of group codes (up to 200 distinct values) associated with treatments. Group codes may be used to modify the A-efficiency measure in the optimisation.

checkGroup
Treatment group code for check treatments.

maxInterchanges
Maximum number of interchanges used in the search to spatially optimise the block design created by the call to nurseryDiGGer.

rngSeeds
Seeds c(s1,s2) to control the DiGGer search. s1 must be in the range [0,31328], s2 must be in the range [0,30081].

Details

nurseryDiGGer creates an optimised block design for the check treatments in an unreplicated design using the blocking sequence specified by the blockSequence parameters Each dimension pair is used in a search within DiGGer. The location of check treatments is optimised using the spatial model specified. The optimised design for check treatments is augmented with the remaining treatments respecting replicate blocks to give an initialDesign of a DiGGer call. An initialSwap matrix is used to keep the check treatments in fixed positions. The DiGGer search is repeated with the full design to optimise the placement of the remaining treatments in the presence of fixed check treatments.

Value

A DiGGer object with the results of a DiGGer search.

Author(s)

Neil Coombes

References

Coombes, N.E. (2002) The Reactive Tabu Search for Efficient Correlated Experimental Designs. PhD Thesis, Liverpool John Moores University.

Examples

# design for 10 entries [6x9] in [6x3] replicates
#  with block sequence [6x1] and [3x1].
# Check varieties 1:2 at 5 repsPerRep
n10 <- nurseryDiGGer(10, 6,9, 6,3,
               blockSeq = list(c(6,1), c(3,1)),
               treatRep = rep(c(5,1),c(2,8)),
               treatGroup = rep(c(2,1), c(2,8)),
                     checkGroup = 2)

[Package DiGGer version 1.0.5 Index]