plot.DiGGer {DiGGer}	R Documentation

Plot DiGGer design labels.

Description

Plot the matrix of DiGGer design numbers from a DiGGer search object using the des.plot function.

Usage

## S3 method for class 'DiGGer'
plot(this, option = "design", bwidth = 3, bcolour = 4, new =
TRUE, ...)

Arguments

this
A DiGGer design object.

option
The only option currently available is "design".

bwidth
Border width for plotting block boundaries.

bcolour
Border colour for plotting block boundaries.

new
Create a new plot (TRUE) or add to existing plot (FALSE).

...
other parameters to des.plot or par.

Details

If a search has not been run, the initial design is plotted. If the search has been run the final design is plotted.

Author(s)

Neil Coombes

[Package DiGGer version 1.0.5 Index]