prDiGGer {DiGGer}	R Documentation

Function to generate a DiGGer search for a p-rep design

Description

Generate a DiGGer search for designs with some unreplicated treatments.

Usage

prDiGGer(numberOfTreatments, rowsInDesign, columnsInDesign,
blockSequence = NULL, betweenRowCorrModel = "AR", betweenRowCorr = 0.5,
betweenColumnCorrModel = "AR", betweenColumnnCorr = 0.5, treatName =
NULL, treatNumber = NULL, treatRepPerRep = NULL, treatGroup = NULL,
maxInterchanges = 20000, targetGroup = 1, runSearch = FALSE, splitOpt =
NULL, rngSeeds = NULL)

Arguments

numberOfTreatments
Number of distinct treatments in the design.

rowsInDesign
Number of rows in the design.

columnsInDesign
Number of columns in the design.

blockSequence
List of dimension pairs of blocks to be optimised in sequence.

betweenRowCorrModel
The correlation pattern between rows may be "AR" AutoRegressive, "MA" Moving Average, or "ID" no correlation.

betweenRowCorr
Valid parameter values p, for "AR" -1<p<1, for "MA" -0.5<p<0.5 and for "ID" p=0.

betweenColumnCorrModel
As for betweenRowCorrModel.

betweenColumnCorr
As for betweenRowCorr.

treatName
Vector of texts giving the names of the treatment.

treatNumber
Vector of numbers associated with the treatments.

treatRepPerRep
Vector of replication levels for each treatment over the whole design.

treatGroup
Vector of group codes (up to 200 distinct values) associated with treatments. Group codes may be used to modify the A-efficiency measure in the optimisation.

maxInterchanges
Maximum number of interchanges used in the search to spatially optimise the block design created by the call to prDiGGer.

targetGroup
Test treatment group used for optimisation in the spatial search phase.

runSearch
Logical value, whether to run the final spatial search immediately.

splitOpt
Dimension pair used for sub-optimal spatial distribution of replicated entries.

rngSeeds
Seeds c(s1,s2) to control the DiGGer search. s1 must be in the range [0,31328], s2 must be in the range [0,30081].

Details

prDiGGer creates an optimised block design for the replicated treatments using the blocking sequence specified by the blockSequence parameters Each dimension pair is used in a search within DiGGer. The optimised block design for replicated treatments is augmented with the unreplicated treatments to give an initialDesign of a DiGGer call. The DiGGer object created has the specifications for a search, using the spatial parameters in the prDiGGer call together with row and column blocks. rowColumn=TRUE, to spatially optimise treatments within the final blocks. The run method must be used to search for the optimised design. For targetGroup=1 the optimisation uses aType="A11" to target comparisons between Group 1 treatments of interest.

Value

A DiGGer object with the specifications for a search or the results of a DiGGer search.

Author(s)

Neil Coombes

References

Coombes, N.E. (2002) The Reactive Tabu Search for Efficient Correlated Experimental Designs. PhD Thesis, Liverpool John Moores University.

Cullis, B.R., Smith, A.B. and Coombes, N.E. (2006) On the design of early generation variety trials with correlated data. JABES 11, 381–393.

Herzberg, A.M. and Jarrett, R.G. (2007). A-optimal block designs with additional singly replicated treatments. Journal of Applied Statistics 34, 61–70.

Examples

# design for 184 entries [23x10] with block sequence
# [23x5], [6x5] and [6x1].
# The first 179 treatments are of interest - Group 1
prep184 <- prDiGGer(numberOfTreatments=184,
                    rowsInDesign=23,
                    columnsInDesign=10,
                    treatRepPerRep=rep(c(2,1,5),c(26,153,5)),
                    treatGroup=rep(c(1,2),c(179,5)),
                    blockSequence=list(c(23,5),c(6,5),c(6,1)))
prep184
# run the search to spatially optimise the design
prep184 <- run(prep184)
# get the matrix and plot it with checks in rep
# duplicated treatments of interest in yellow.
m184 <- getDesign(prep184)
desPlot(m184,seq(153)+26,col=8,new=TRUE,label=FALSE)
desPlot(m184,seq(26),col=7,new=FALSE,label=TRUE)
desPlot(m184,seq(5)+179,col=2,new=FALSE,label=TRUE,
         bdef=cbind(6,1),bcol=4,bwd=4)

[Package DiGGer version 1.0.5 Index]