print.Block {DiGGer}	R Documentation

Print a Block object.

Description

Print a Block object.

Usage

## S3 method for class 'Block'
print(this, blockid = "", type = "description", RowsInDesign =
NULL, ColumnsInDesign = NULL, ...)

Arguments

this
A DiGGer Bock object.

blockid
A block number to be printed.

type
Print a type="description" of the Block object or type="list" produce a listing of block number for each unit in the design.

RowsInDesign
Number of rows in the design, needed for type="list".

ColumnsInDesign
Number of columns in the design, needed for type="list".

...
Not used.

Details

Prints a description of the Block object to the terminal screen when type="description". Creates a column vector of block numbers when type="list" and RowsInDesign and ColumnsInDesign are supplied.

Author(s)

Neil Coombes

[Package DiGGer version 1.0.5 Index]