print.DiGGer {DiGGer}	R Documentation

Print a DiGGer search object.

Description

Print a DiGGer search object.

Usage

## S3 method for class 'DiGGer'
print(this, option = "search", file = NULL, ...)

Arguments

this
A DiGGer search object.

option
The default option="search" prints a summary of the planned search to the screen if file=NULL or prints to a file in the format required to run DiGGer externally. Other options print: the initial design with option="idsgn", the initial swap codes with option="swap", the treatment information details with option="treatment", and the field book listing of the final design with option="list".

file
The file basename to print to.

...
Not used.

Details

Printing a DiGGer object with option="search" to the computer screen gives a schematic representation of the DiGGer search.

Search initialisation information precedes the information for each Phase call to the DiGGer program. This includes the number and type of treatments, the dimensions of the design and the dimensions of each replicate within the design.

The Phase header information shows the dimensions of swap and correlation template blocks, the DiGGer stopping rules, target A-measure and maximum interchanges, and the type of A-measure and search intensity.

Each phase has at least one Objective and the weight, linear covariate, spatial model and blocking structures are summarised.

With option="idsgn" the matrix representation of the initial design is printed.

With option="iswap" the matrix of initial swap codes is printed.

With option="treatment" the data frame of treatment names, treatment numbers, treatment repeats per replicate and treatment group codes is printed.

option="list" should only be used once the DiGGer search has been run. This produces a field-book listing of the design, rows nested within columns, with treatment and design information.

option="all" prints all of the above.

Printing to file produces the files needed for DiGGer to be run from a command window using the standalone program.

Author(s)

Neil Coombes

[Package DiGGer version 1.0.5 Index]