r2dDiGGer {DiGGer}	R Documentation

Function for strict two dimensional replicates from a DiGGer design search.

Description

r2dDiGGer attempts to form replicates in two directions in a DiGGer search. Replicates in the second direction may not be rectangular.

Usage

r2dDiGGer(numberOfTreatments, rowsInDesign, columnsInDesign,
rowsInReplicate = NULL, columnsInReplicate = NULL,
maxInterchanges = 1e+05, searchIntensity = 100,
aType = NULL, runSearch = TRUE, rngSeeds = NULL,
treatName = NULL, treatNumber = NULL, treatRepPerRep = NULL,
treatGroup = NULL, spatial = TRUE, rowColumn = TRUE)

Arguments

numberOfTreatments
The number of treatments in the design.

rowsInDesign
The number of rows in the design.

columnsInDesign
The number of columns in the design.

rowsInReplicate
The number or rows in the template replicate block.

columnsInReplicate
The number of columns in the template replicate block.

maxInterchanges
Number of treatment interchanges to test in each search phase.

searchIntensity
Percentage of possible interchanges to consider for non-improving interchanges.

aType
A DiGGer A-measure type: "A++", "Agg", "A22", "A11", "A1+", "Aa2" or "Aaa".

runSearch
Logical value, whether to run the search immediately.

rngSeeds
Seeds c(s1,s2) to control the DiGGer search. s1 must be in the range [0,31328], s2 must be in the range [0,30081].

treatName
Vector of treatment names to be associated with the design numbers.

treatNumber
Vector of treatment numbers to be associated with the design numbers.

treatRepPerRep
Vector of replication levels for each treatment within each replicate template block.

treatGroup
Vector of group codes (up to 200 distinct values) associated with treatments. Group codes may be used to modify the A-efficiency measure in the optimisation.

spatial
If TRUE separable AR processes between rows and between columns are specified in the final DiGGer search phase.

rowColumn
If TRUE random rows, random columns and random error are specified in the final DiGGer search phase.

Details

r2dDiGGer attempts to create a design that is strictly replicated in two directions. The second set of replicates may not be rectangular. Treatments are arranged within the sub-blocks to optimise the design with respect to the spatial and rowColumn parameters specified.

Note that two replicate designs will not be replicated in two directions as sub-blocks are disconnected in this case.

If there is more than one treatGroup code the search may optimise comparisons between treatments in different treatment groups, aType="Agg". If aType=NULL the default code for equally replicated treatments is "A++" and for unequal replication it is "Agg".

The search to replicate in two directions is run immediately, the final row-column and /or spatial search is run immediately if runSearch=TRUE. Otherwise the final search may be modified using set, add and delete methods associated with the search parameters and the run function then used to run the search.

Value

A DiGGer search object.

ddphase
A list with an element for each phase of the search. An element is populated with seeds, design, aMeasures and lastImprovement as each search phase is completed.

dlist
A dataframe of the design after a search has been run().

Author(s)

Neil Coombes

References

Coombes, N.E. (2002) The Reactive Tabu Search for Efficient Correlated Experimental Designs. PhD Thesis, Liverpool John Moores University.

Examples

# Example of search run for 20 treatments
# arranged in 20 rows by 3 columns
# with replicates 20 rows by 1 column
d20 <- r2dDiGGer(20, 20,3, 20,1)

[Package DiGGer version 1.0.5 Index]