run {DiGGer}	R Documentation

Run a DiGGer search.

Description

Run a DiGGer search to optimise a design subject to specified correlation, blocking and A-measure type.

Usage

## S3 method for class 'DiGGer'
run(this, newStart = FALSE, newPath = FALSE, err =
errcodes, continue=FALSE, ...)

Arguments

this
A DiGGer object.

newStart
If newStart is TRUE create a new starting design else use the current starting design.

newPath
If newPath is TRUE the random number generator is re-initialised.

err
Data file of error codes.

continue
If TRUE a search will be run even if the DiGGer object contains a completed search, otherwise a search will not be run to prevent overwriting the current search.

...
Not used.

Value

A DiGGer design object.

ddphase
A list with an element for each phase of the search initially set NULL. An element is populated with seeds, design, aMeasures and lastImprovement as each search phase is completed.

dlist
A data frame holding the field-book listing of the design, with rows nested within columns of the design.

errcode
The numeric code for the error status of the DiGGer run. Non-zero codes indicate a fault in the search.

idsgn
The matrix of design numbers used as the InitialDesign in the search.

iswap
The matrix of swap codes used to restrict interchanges between experimental units of the design.

treatment
A dataframe with columns Name, Number, Repeats and Group detailing the treatment specifications for the design.

Author(s)

Neil Coombes

References

Coombes, N.E. (2002) The Reactive Tabu Search for Efficient Correlated Experimental Designs. PhD Thesis, Liverpool John Moores University.

Examples

# Example of search set up for 21 treatments
# arranged in 21 rows by 3 columns
# with replicates 21 rows by 1 column
d21 <- corDiGGer(21, 21,3, 21,1)
d21 <- run(d21)

[Package DiGGer version 1.0.5 Index]