setAType {DiGGer}	R Documentation

Set the type of A-Measure for a DiGGer search phase.

Description

Set the A-measure type for a phase of a DiGGer search.

Usage

## S3 method for class 'DiGGer'
setAType(this, phaseNo = 1, aType = "A++", ...)
## S3 method for class 'Phase'
setAType(this, aType = "A++", ...)

Arguments

this
A DiGGer search object.

phaseNo
The Phase number to be modified.

AType
A DiGGer A-measure type: "A++", "Agg", "A22", "A11", "A1+", "Aa2" or "Aaa".

...
Not used.

Details

DiGGer optimisation is based on the average variance of treatment contrasts. The default, aType="A++", optimises the average variance of pairwise contrasts amongst all treatments. Other A-measure types are based on the treatment group code defined for each treatment in the design.

If aType="Agg" the average variance between all pairs of treatments from different treatment groups is minimised.

If aType="A22" the average variance of pairwise comparisons of treatments in group 2 is optimised.

If aType="A11" the average variance of pairwise comparisons of treatments in group 1 is optimised.

If aType="A1+" the average variance of pairwise comparisons of treatments in group 1 with treatments in other groups is optimised.

If aType="Aa2" the comparisons optimised are between the average of group 1 treatments and each treatment in group 2.

If aType="Aaa" the comparison optimised is the average of group 1 treatments with the average of group 2 treatments.

Author(s)

Neil Coombes

[Package DiGGer version 1.0.5 Index]