setBlock {DiGGer}	R Documentation

Modify the parameters for a Block object.

Description

Modify the number of rows and the number of columns in the template of a DiGGer Block. Define the random block variance ratio or set to a non-positive value for fixed blocks. The Block object must already exist.

Usage

## S3 method for class 'DiGGer'
setBlock(this, phaseNo = 1, objectiveNo = 1, blockNo = 1,
rowsInBlock = 1, columnsInBlock = 1, blockVarianceRatio = 0.1, ...)
## S3 method for class 'Phase'
setBlock(this, objectiveNo = 1, blockNo = 1,
rowsInBlock = 1, columnsInBlock = 1, blockVarianceRatio = 0.1, ...)
## S3 method for class 'Objective'
setBlock(this, blockNo = 1,
rowsInBlock = 1, columnsInBlock = 1, blockVarianceRatio = 0.1, ...)

Arguments

this
The DiGGer object to be modified.

phaseNo
The Phase number of the DiGGer search to be modified.

objectiveNo
The Objective number to be modified.

blockNo
The Block number to be modified.

rowsInBlock
Number of rows in the block template.

columnsInBlock
Number of columns in the block template.

blockVarianceRatio
The block variance ratio, positive for random blocks, fixed blocks otherwise.

...
Not used.

Author(s)

Neil Coombes

See Also

Block

[Package DiGGer version 1.0.5 Index]