setInitialDesign {DiGGer}	R Documentation

Set an Initial Design for a DiGGer search.

Description

Use a supplied design or use the DiGGer search parameters to create a design to use as the starting design in a DiGGer search.

Usage

## S3 method for class 'DiGGer'
setInitialDesign(this, design = NULL, ...)

Arguments

this
A DiGGer search object.

design
If supplied design must be a matrix of design numbers conforming to the design specifications of the DiGGer search.

...
Not used.

Author(s)

Neil Coombes

[Package DiGGer version 1.0.5 Index]