setLinearCovariate {DiGGer}	R Documentation

Set the Linear Covariate in a DiGGer search objective.

Description

Set the Linear Covariate in a DiGGer search objective.

Usage

## S3 method for class 'DiGGer'
setLinearCovariate(this, phaseNo = 1, objectiveNo = 1,
linearCovariate = "NONE", ...)

## S3 method for class 'Phase'
setLinearCovariate(this, objectiveNo = 1,
linearCovariate = "NONE", ...)

## S3 method for class 'Objective'
setLinearCovariate(this, linearCovariate = "NONE", ...)

Arguments

this
A DiGGer search object.

phaseNo
The phase number of the DiGGer search to be modified.

objectiveNo
The objective number of the DiGGer search phase to be modified.

linearCovariate
A text specifying the linear covariate: "NONE" for no linear covariate, "ROWS" for a linear covariate based on row number, "COLUMNS" for a linear covariate based on column number or "BOTH" for two covariates, one for rows and one for columns in the design.

...
Not used.

Author(s)

Neil Coombes

[Package DiGGer version 1.0.5 Index]