setMaxInterchanges {DiGGer}	R Documentation

Set the maximum number of treatment interchanges in a DiGGer search Phase.

Description

Set the maximum number of treatment interchanges in a DiGGer search Phase.

Usage

## S3 method for class 'DiGGer'
setMaxInterchanges(this, phaseNo = 1, maxInterchanges = 1e+05,
...)
## S3 method for class 'Phase'
setMaxInterchanges(this, maxInterchanges = 1e+05, ...)

Arguments

this
A DiGGer search object.

phaseNo
The phase number of the DiGGer search to be modified.

maxInterchanges
The maximum number of interchanges to test in this search Phase.

...
Not used.

Author(s)

Neil Coombes

[Package DiGGer version 1.0.5 Index]