setPhase {DiGGer}	R Documentation

Modify a DiGGer search Phase.

Description

Modify an existing DiGGer search Phase.

Usage

## S3 method for class 'DiGGer'
setPhase(this, phaseNo = 1, ...)

Arguments

this
A DiGGer search object.

phaseNo
The Phase number of the DiGGer object to be modified.

...
A DiGGer Phase() object.

Details

A DiGGer Phase object controls a search within DiGGer. The initial phase will use the initial design supplied to, or constructed by, the DiGGer() constructor function. Subsequent search phases use the best design found in the previous search phase.

The search Phase defines the template blocks for allowable swaps and for blocks of correlated error. The A-measure type, target A-measure, maximum treatment interchanges to test and search intensity paramaters are specified. The number of objectives is given numberOfObjectives >= 1, and an Objective object supplied for each of these objectives.

Author(s)

Neil Coombes

[Package DiGGer version 1.0.5 Index]