setSearchIntensity {DiGGer}	R Documentation

Set the Search Intensity for a DiGGer search phase.

Description

Modify the Search Intensity parameter for a DiGGer search phase to control the percentage of possible interchanges to consider for a non-improving treatment interchange.

Usage

## S3 method for class 'DiGGer'
setSearchIntensity(this, phaseNo = 1, searchIntensity = 100, ...)
## S3 method for class 'Phase'
setSearchIntensity(this, searchIntensity = 100, ...)

Arguments

this
A DiGGer search object.

phaseNo
The DiGGer Phase number to be modified.

searchIntensity
A value [0,100], representing the percentage of interchanges to consider for a non-improving treatment interchange. If searchIntensity=0 the first valid randomly chosen treatment interchange is used.

...
Not used.

Details

The DiGGer search uses a randomisation of all possible treatment plot interchanges to find an improvement to the current design in the search. If no improvement is found after all interchanges have been tested the searchIntensity defines the percentage of interchanges to consider for a non-improving move. Low search intensities are useful for finding optimal designs with a single block factor.

Author(s)

Neil Coombes

[Package DiGGer version 1.0.5 Index]