setSwapBlock {DiGGer}	R Documentation

Set the template for the swap blocks in a DiGGer search Phase.

Description

Set the template for the swap blocks in a DiGGer search Phase.

Usage

## S3 method for class 'DiGGer'
setSwapBlock(this, phaseNo = 1, rowsInSwapBlock,
columnsInSwapBlock, ...)
## S3 method for class 'Phase'
setSwapBlock(this, rowsInSwapBlock,
columnsInSwapBlock, ...)

Arguments

this
A DiGGer search object.

phaseNo
The Phase number to be modified.

rowsInSwapBlock
Number of rows in the swap block template.

columnsInSwapBlock
Number of columns in the swap block template.

...
Not used.

Details

Treatment interchanges are allowed in the DiGGer search between plots in the same swap block having the same swap code. Swap blocks are determined by tiling the template from Row 1, Column 1 down and then across the design. Unless initiial swap codes are supplied all experimental units have the same swap code.

Author(s)

Neil Coombes

[Package DiGGer version 1.0.5 Index]