setTargetAValue {DiGGer}	R Documentation

Set the target A-value.

Description

Set the A-value to be used as a stopping rule in a DiGGer search Phase. A value of 0.0 is unattainable so this default ignores the A-value efficiency measure as a stopping rule in the search.

Usage

## S3 method for class 'DiGGer'
setTargetAValue(this, phaseNo = 1, targetAValue = 0, ...)
## S3 method for class 'Phase'
setTargetAValue(this, targetAValue = 0, ...)

Arguments

this
A DiGGer object.

phaseNo
The Phase number to be modified.

targetAValue
The A-value used to stop a DiGGer search phase. When the current A-value falls below the targetAValue the current search phase is stopped.

...
Not used.

Author(s)

Neil Coombes

[Package DiGGer version 1.0.5 Index]