setTreatmentVarianceRatio {DiGGer}	R Documentation

Set the treatment variance ratio.

Description

For treatments considered random the treatment variance ratio is positive, a zero or negative treatment variance ratio causes treatments to be considered fixed, the default.

Usage

## S3 method for class 'DiGGer'
setTreatmentVarianceRatio(this, treatmentVarianceRatio = -1, ...)

Arguments

this
A DiGGer object.

treatmentVarianceRatio
The treatment variance ratio, if treatmentVarianceRatio>0 treatments are considered random, otherwise treatments are considered fixed.

...
Not used.

Author(s)

Neil Coombes

[Package DiGGer version 1.0.5 Index]