summary {DiGGer}	R Documentation

Summarise the DiGGer search.

Description

Produce a summary of the DiGGer search, giving a schematic of the search phases, details of the final A-value and the interchange at this final improvement of each search phase and a suggested random model to be used in the analysis.

Usage

## S3 method for class 'DiGGer'
summary(this, ...)

Arguments

this
A DiGGer object.

...
Not used.

Author(s)

Neil Coombes

[Package DiGGer version 1.0.5 Index]